llm-assistant/pyproject.toml
[project]
name = "llm-assistant"
version = "1.0.0"
description = "Terminal AI Assistant for Terminator"
requires-python = ">=3.10"
dependencies = [
    "llm>=0.27",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "pydantic>=2.0",
    "jinja2>=3.0",
    "PyYAML>=6.0",
    "iso639-lang>=2.0",
]

[project.optional-dependencies]
voice = ["sounddevice", "numpy", "onnx-asr"]
tts = ["google-cloud-texttospeech", "google-auth", "strip-markdown"]
web = ["fastapi", "uvicorn"]
clipboard = ["pyperclip"]
all = ["llm-assistant[voice,tts,web,clipboard]"]

[project.scripts]
llm-assistant = "llm_assistant.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["llm_assistant"]
